Fix Small Bugs (#110)
* Fixes underflow bug and wrongly-named accessories command

* weird spacing in CanSensorBms

+++ src/Sensor/CanSensorOrionBms.cpp
@@ -121,10 +121,10 @@ void CanSensorOrionBms::update(CanMessage message) {
 			_validationMap[CAN_ORIONBMS_CELL] = _lastUpdateTime;
 			break;
 		case CAN_ORIONBMS_TEMP:
-			_batteryTempMin = (int)message.data[0];
-			_batteryTempMax = (int)message.data[1];
-			_batteryTempAvg = (int)message.data[2];
-			_tempBms = (int)message.data[3];
+			_batteryTempMin = (int8_t)message.data[0];
+			_batteryTempMax = (int8_t)message.data[1];
+			_batteryTempAvg = (int8_t)message.data[2];
+			_tempBms = (int8_t)message.data[3];
 			_validationMap[CAN_ORIONBMS_TEMP] = _lastUpdateTime;
 			break;
 		default:
